Fix texture extraction for Chaos Bleeds (Xbox)
Reading the data size for Chaos Bleeds textures seemingly always returns a value of 88, which is obviously incorrect. The calculated size **must** be used instead. This adds an additional condition for Chaos Bleeds (`!(version == 170 && platform == Platform::Xbox)`) when deciding whether the `data_size` field should be included in `EXGeoTexture`. Disabling this field for Chaos Bleeds allows textures to be properly extracted.